requirements.txt
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
Flask==3.0.3
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pymongo==4.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.0.4
WTForms==3.2.1
